auth-solution/requirements.txt
flask==2.2.2
Flask-SQLAlchemy== 3.0.3
flask-marshmallow == 0.14.0
redis == 4.5.1
flask-jwt-extended == 4.4.4
flask-security == 3.0.0
email-validator==1.3.1
gevent==22.10.2
pydantic==1.9.0
psycopg2==2.9.5
gunicorn==20.1.0
flask-migrate==4.0.4
pytz==2022.7.1
opentelemetry-api==1.16.0
opentelemetry-instrumentation-flask==0.37b0
opentelemetry-exporter-jaeger==1.12.0
user-agents==2.2.0
rauth==0.7.3
Flask-reCaptcha==0.4.2
user-agents==2.2.0